Support explicit job-name
GitHub only exposes `github.job` (the Job Id) or, if there was an
explicit name given, you can do `github.jobs[github.job].name`. There is
no way to get the complete Job Name given in the case of a matrix Job.

This means that, for example, we can't look up the Job URL for `deploy
(prod)` because we only know the `deploy` part.

The various actions that do this all accept an explicit name input, so
that you can pass it from the outside where you can include `${{
matrix.env }}` yourself -- so we support that here now.