Include the unknown author in the message body
Originally, we assumed that seeing a user avatar in the notification,
but an `<@unknown>` in the message body would directly indicate that
that's the user that needs to be added to the slack-users json.

However, there are circumstances where that's not the case. Notably,
where the PR author or workflow actor isn't the same person that made
the commit. This has led to some cases where _PR_ authors were asked to
address issues in the slack-users json when there were none. It was just
that the _commit_ author was a bot.

Showing the commit author in the message will prevent this confusion.